Moved all audio rating code to a single documented example
